Extracted IODevice component
