+++ HW2/Vladimir_Chirushin/doors-levels/IInputOutputDevice.cs
@@ -1,0 +1,8 @@
+﻿namespace doors_levels
+{ 
+        public interface IInputOutputDevice
+        {
+            string ReadInput();
+            void WriteOutput(string dataToOutput);
+        }
+}
