+++ HW2/Vladimir_Chirushin/doors-levels/doors-levels.csproj
@@ -43,6 +43,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ConsoleIODevice.cs" />
+    <Compile Include="DoorsGame.cs" />
+    <Compile Include="IInputOutputDevice.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
